backend/requirements.txt
# FastAPI + ASGI
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Google Gemini
google-generativeai>=0.7.0

# LangGraph + LangChain core
langgraph>=0.1.0
langchain-core>=0.2.0

# Supabase
supabase>=2.4.0

# Data ingestion
yfinance>=0.2.40
python-epo-ops-client>=3.1.0

# Web scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Utilities
python-dotenv>=1.0.0
